Disable strict-type-predicates tslint rule
If I declare an expression as number | null, then if I do a compare:

let v: number | null;
if (v == null) {
}

strict-type-predicates fails because type of v can not be undefined,
tslint recommend to use '===' instead of '=='.

While type is one-thing, runtime is another, it may has a chance to be
undefined.